Fix net/http requirement
In a previous commit, I removed this requirement because I thought it wasn't necessary. But it was, I jsut hadn't restarted the server before trying it out (this is why we should have tests!)

so I've added it back in.

I also removed the rescue, as now this feature is working, I want it to error properly if problems.